LayoutTests/imported/w3c/web-platform-tests/workers/semantics/multiple-workers/003-expected.txt

CONSOLE MESSAGE: ReferenceError: Can't find variable: Worker

FAIL creating 3 nested dedicated workers assert_unreached: error Reached unreachable code
